ffmpeg_sys_next

Function av_dict_iterate

pub unsafe extern "C" fn av_dict_iterate(
    m: *const AVDictionary,
    prev: *const AVDictionaryEntry,
) -> *const AVDictionaryEntry

Iterate over a dictionary

Iterates through all entries in the dictionary.

@warning The returned AVDictionaryEntry key/value must not be changed.

@warning As av_dict_set() invalidates all previous entries returned by this function, it must not be called while iterating over the dict.

Typical usage: @code const AVDictionaryEntry *e = NULL; while ((e = av_dict_iterate(m, e))) { // … } @endcode

@param m The dictionary to iterate over @param prev Pointer to the previous AVDictionaryEntry, NULL initially

@retval AVDictionaryEntry* The next element in the dictionary @retval NULL No more elements in the dictionary