ffmpeg_sys_next

Function avcodec_get_supported_config

pub unsafe extern "C" fn avcodec_get_supported_config(
    avctx: *const AVCodecContext,
    codec: *const AVCodec,
    config: AVCodecConfig,
    flags: c_uint,
    out_configs: *mut *const c_void,
    out_num_configs: *mut c_int,
) -> c_int

Retrieve a list of all supported values for a given configuration type.

@param avctx An optional context to use. Values such as strict_std_compliance may affect the result. If NULL, default values are used. @param codec The codec to query, or NULL to use avctx->codec. @param config The configuration to query. @param flags Currently unused; should be set to zero. @param out_configs On success, set to a list of configurations, terminated by a config-specific terminator, or NULL if all possible values are supported. @param out_num_configs On success, set to the number of elements in out_configs, excluding the terminator. Optional.