ffmpeg_sys_next

Function avformat_write_header

pub unsafe extern "C" fn avformat_write_header(
    s: *mut AVFormatContext,
    options: *mut *mut AVDictionary,
) -> c_int

Allocate the stream private data and write the stream header to an output media file.

@param s Media file handle, must be allocated with avformat_alloc_context(). Its \ref AVFormatContext.oformat “oformat” field must be set to the desired output format; Its \ref AVFormatContext.pb “pb” field must be set to an already opened ::AVIOContext. @param options An ::AVDictionary filled with AVFormatContext and muxer-private options. On return this parameter will be destroyed and replaced with a dict containing options that were not found. May be NULL.

@retval AVSTREAM_INIT_IN_WRITE_HEADER On success, if the codec had not already been fully initialized in avformat_init_output(). @retval AVSTREAM_INIT_IN_INIT_OUTPUT On success, if the codec had already been fully initialized in avformat_init_output(). @retval AVERROR A negative AVERROR on failure.

@see av_opt_find, av_dict_set, avio_open, av_oformat_next, avformat_init_output.